refactor: queryAllTags to a single file
