--- packages/core/src/domains/tag/queryAllTags/query.ts
@@ -1,7 +1,0 @@
-import groq from 'groq';
-
-export const allTagsQuery = groq`
-*[_type == "tag" && !(_id in path('drafts.**'))] | order(slug.current asc){
-  "slug": slug.current
-}
-`;
